requirements/test.txt
-r base.txt

pytest-django==3.4.3
freezegun==0.3.11
lxml==4.2.5
